// src/api/orval/model/enumsCertExpiryRunStatus.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */

export type EnumsCertExpiryRunStatus = typeof EnumsCertExpiryRunStatus[keyof typeof EnumsCertExpiryRunStatus];


export const EnumsCertExpiryRunStatus = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
} as const;